requirements.txt
beautifulsoup4==4.12.3
Flask==3.0.2
gunicorn==21.2.0
psycopg[binary]==3.1.18
python-dateutil==2.9.0.post0
requests==2.31.0
SQLAlchemy==2.0.29
